requirements.txt
requests~=2.32.4
python-telegram-bot~=13.15
openai~=1.98.0
peewee~=3.18.2
python-dotenv~=1.1.1
pyTelegramBotAPI~=4.27.0